--- src/app/lib/placeholder.ts
@@ -1,5 +1,0 @@
-export const emailTemplates = [
-  { name: "template 1", body: "test body 1" },
-  { name: "template 2", body: "test body 2" },
-  { name: "template 3", body: "test body 3" },
-];
